Authentication is developed.

+++ app/Http/Controllers/QuestionsController.php
@@ -63,6 +63,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        // Authentication
+        if(\Gate::denies('update-question', $question)){
+            abort(403, 'Access Denied!');  
+        }
         return view('questions.edit', compact('question')); 
     }
 
@@ -75,6 +79,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        // Authentication 
+        if(\Gate::denies('update-question', $question)){
+            abort(403, 'Access Denied!');  
+        }
+
         // Validation data
         $data = $request->only('title', 'body');
 
@@ -93,6 +102,11 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        // Authentication
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, 'Access Denied!');  
+        }
+
         $question->delete(); 
         return redirect('/questions')->with('success', 'Your question has been deleted.'); 
     }
